pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "one-piece-manga-downloader"
description = 'Download One Piece Manga Chapters from TCB Scans'
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = []
authors = [
  { name = "Guillem96", email = "guillem.orellana@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "httpx",
  "beautifulsoup4",
  "toml",
  "Pillow",
  "typer"
]
dynamic = ["version"]

[project.scripts]
one-piece-download = "op_downloader.__main__:app"

[project.urls]
Documentation = "https://github.com/Guillem96/one-piece-manga-downloader#readme"
Issues = "https://github.com/Guillem96/one-piece-manga-downloader/issues"
Source = "https://github.com/Guillem96/one-piece-manga-downloader"

[tool.hatch.version]
path = "op_downloader/__version__.py"

[tool.hatch.envs.default]
python = "3.9"
dependencies = [
  "yapf",
  "isort",
  "mypy",
]

[tool.hatch.envs.telegram]
python = "3.9"
dependencies = [
  "yapf",
  "isort",
  "mypy",
  "python-telegram-bot",
  "aws-lambda-typing",
  "typing_extensions"
]

